requirements.txt
requests==2.31.0
beautifulsoup4>=4.9.3
dnspython>=2.2.0
argparse==1.4.0
colorama>=0.4.4
tqdm>=4.62.3
fake-useragent>=0.1.11
aiohttp>=3.8.0
python-wappalyzer>=0.3.1
python-nmap==0.7.1
waybackpy==3.0.6
vulners>=2.0.7
socid-extractor==0.0.24
httpx>=0.23.0
async-timeout==4.0.3
